+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+### Fixed
+
+* Values containing carriage returns are now correctly escaped when using the `stringify` method.
+
 ### Changed
 
 * Moved `stringify` and `parse` methods into their own files.